package parser!

+++ src/com/microsoft/projecta/tools/DeviceConnection.java
@@ -6,6 +6,8 @@ import java.io.BufferedOutputStream;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
 import java.util.Enumeration;
 import java.util.logging.Level;
 import java.util.logging.Logger;
@@ -19,22 +21,23 @@ import com.microsoft.projecta.tools.workflow.WorkFlowSingleProcStage;
 import com.microsoft.projecta.tools.workflow.WorkFlowStatus;
 
 public final class DeviceConnection extends WorkFlowSingleProcStage {
-    private static Logger logger = Logger.getLogger(DeviceConnection.class
-            .getSimpleName());
+    private static Logger logger = Logger.getLogger(DeviceConnection.class.getSimpleName());
     private static final int UNZIP_BUFFER = 2048;
 
-    public static void unZipAll(File source, File destination) throws ZipException, IOException
-    {
-        logger.info("Unzipping - " + source.getName());
-        logger.entering("DeviceConnection", "unZipAll");
+    public void unZipAll(File zippedSdk, File unzippedSdkDir) throws ZipException, IOException {
+        unZipAll(zippedSdk, unzippedSdkDir, false);
+    }
+
+    public void unZipAll(File source, File destination, boolean recursively) throws ZipException,
+            IOException {
+        fireOnLogOutput("Unzipping " + source.getName());
         ZipFile zip = new ZipFile(source);
 
         destination.getParentFile().mkdirs();
         Enumeration<? extends ZipEntry> zipFileEntries = zip.entries();
 
         // Process each entry
-        while (zipFileEntries.hasMoreElements())
-        {
+        while (zipFileEntries.hasMoreElements()) {
             // grab a zip file entry
             ZipEntry entry = (ZipEntry) zipFileEntries.nextElement();
             String currentEntry = entry.getName();
@@ -45,8 +48,7 @@ public final class DeviceConnection extends WorkFlowSingleProcStage {
             // create the parent directory structure if needed
             destinationParent.mkdirs();
 
-            if (!entry.isDirectory())
-            {
+            if (!entry.isDirectory()) {
                 BufferedInputStream is = new BufferedInputStream(zip.getInputStream(entry));
                 int currentByte;
                 // establish buffer for writing file
@@ -60,28 +62,26 @@ public final class DeviceConnection extends WorkFlowSingleProcStage {
                 while ((currentByte = is.read(data, 0, UNZIP_BUFFER)) != -1) {
                     dest.write(data, 0, currentByte);
                 }
-                logger.log(Level.INFO, "unzipped "+entry.getName());
+                fireOnLogOutput("Unzipped " + entry.getName());
                 dest.close();
                 fos.close();
                 is.close();
             } else {
                 // Create directory
                 destFile.mkdirs();
-                logger.log(Level.INFO, "creating "+destFile.getAbsolutePath());
+                fireOnLogOutput("Creating " + destFile.getAbsolutePath());
             }
 
-            if (currentEntry.endsWith(".zip"))
-            {
+            if (recursively && currentEntry.endsWith(".zip")) {
                 // found a zip file, try to unzip it as well
                 unZipAll(destFile, destinationParent);
                 // delete the unzipped file
                 if (!destFile.delete()) {
-                    logger.log(Level.WARNING, "Could not delete zip");
+                    fireOnLogOutput(logger, Level.WARNING, "Creating " + destFile.getAbsolutePath());
                 }
             }
         }
         zip.close();
-        logger.exiting("DeviceConnection", "unZipAll");
     }
 
     private LaunchConfig mConfig;
@@ -95,26 +95,26 @@ public final class DeviceConnection extends WorkFlowSingleProcStage {
     public WorkFlowStatus getStatus() {
         return WorkFlowStatus.DEVICE_CONNECTED;
     }
-    
-    private File getOsSubdirZipFile(String sdkToolsPath) {
-        File subdirZip = null;
+
+    private Path getOsSubdirZipFile(String sdkToolsPath) {
+        Path subdirZip = null;
         if (OS.CurrentOS == OS.WINDOWS) {
-            subdirZip = path(sdkToolsPath, "PC", "ProjectA-windows.zip").toFile();
+            subdirZip = path(sdkToolsPath, "PC", "ProjectA-windows.zip");
         } else if (OS.CurrentOS == OS.MAC) {
-            subdirZip = path(sdkToolsPath, "MAC", "ProjectA-darwin.zip").toFile();
+            subdirZip = path(sdkToolsPath, "MAC", "ProjectA-darwin.zip");
         } else {
-            failfast("Not supported platform: "+OS.CurrentOS);
+            failfast("Not supported platform: " + OS.CurrentOS);
         }
         return subdirZip;
     }
-    
-    private boolean isSdkBuildDrop(String sdkToolsPath){
-        File subdirTools = path(sdkToolsPath, "tools").toFile();
-        if (subdirTools.exists() && subdirTools.isDirectory()) {
+
+    private boolean isSdkBuildDrop(String sdkToolsPath) {
+        Path subdirToolsDir = path(sdkToolsPath, "tools");
+        if (Files.exists(subdirToolsDir) && Files.isDirectory(subdirToolsDir)) {
             return false;
         } else {
-            File subdirOs = getOsSubdirZipFile(sdkToolsPath);
-            if (subdirOs.exists() && subdirOs.isFile()) {
+            Path subdirOsZip = getOsSubdirZipFile(sdkToolsPath);
+            if (Files.exists(subdirOsZip) && Files.isRegularFile(subdirOsZip)) {
                 return true;
             }
         }
@@ -130,26 +130,25 @@ public final class DeviceConnection extends WorkFlowSingleProcStage {
             mConfig.setUnzippedSdkToolsPath(mConfig.getSdkToolsPath());
             setup_result = true;
         } else {
-            File zippedSdk = getOsSubdirZipFile(mConfig.getSdkToolsPath());
-            File unzippedSdk = path(mConfig.getOutdirPath(), "sdkTools").toFile();
-            if (unzippedSdk.exists() && unzippedSdk.isDirectory()) {
+            Path zippedSdk = getOsSubdirZipFile(mConfig.getSdkToolsPath());
+            Path unzippedSdkDir = path(mConfig.getOutdirPath(), "sdkTools");
+            if (Files.exists(unzippedSdkDir) && Files.isDirectory(unzippedSdkDir)) {
                 // TODO check if this is the same version as mConfig.getSdkToolsPath()
-                mConfig.setUnzippedSdkToolsPath(unzippedSdk.getAbsolutePath());
+                mConfig.setUnzippedSdkToolsPath(unzippedSdkDir.toAbsolutePath().toString());
                 setup_result = true;
             } else {
                 // Unzip
                 try {
-                    unZipAll(zippedSdk, unzippedSdk);
+                    unZipAll(zippedSdk.toFile(), unzippedSdkDir.toFile());
+                    mConfig.setUnzippedSdkToolsPath(unzippedSdkDir.toAbsolutePath().toString());
                     setup_result = true;
                 } catch (IOException e) {
                     // TODO clean up the unzipped folder?
                     setup_result = false;
-                    String error_msg = String.format(
-                            "Error occurred while unzipping sdk tools from %s to %s",
-                            zippedSdk.getAbsolutePath(), unzippedSdk.getAbsolutePath());
-                    fireOnLogOutput(error_msg);
-                    fireOnLogOutput(e.getMessage());
-                    logger.log(Level.SEVERE, error_msg, e);
+                    fireOnLogOutput(logger, Level.SEVERE, String.format(
+                            "Error occurred while unzipping sdk tools from %s to %s", zippedSdk
+                                    .toAbsolutePath().toString(), unzippedSdkDir.toAbsolutePath()
+                                    .toString()), e);
                 }
             }
         }
@@ -164,8 +163,7 @@ public final class DeviceConnection extends WorkFlowSingleProcStage {
         // TODO save the log somewhere?
         return new ProcessBuilder().command(
                 join(mConfig.getUnzippedSdkToolsPath(), "tools", "wconnect.exe"),
-                mConfig.getDeviceIPAddr()).directory(
-                new File(mConfig.getOutdirPath()));
+                mConfig.getDeviceIPAddr()).directory(new File(mConfig.getOutdirPath()));
     }
 
 }
